injectables renamed to services
