+++ packages/seed/src/config.ts
