+++ packages/seed/src/index.ts
@@ -1,3 +1,4 @@
 export { Fixture, createResponse } from './fixture'
 export { LocalConfig } from './localConfig'
 export { FixtureStorage } from './fixtureStorage'
+export { config } from './config'
