+++ packages/services-http/seed/https/postman-echo.com/get?foo1=bar1&foo2=bar2/default.json
