+++ packages/services-http/test/module.ts
@@ -1,10 +1,10 @@
 import { Mode, SetMode } from '@seagull/mode'
+import { config } from '@seagull/seed'
 import { BasicTest } from '@seagull/testing'
 import { expect } from 'chai'
 import 'chai/register-should'
 import { Container } from 'inversify'
 import { skip, slow, suite, test, timeout } from 'mocha-typescript'
-import { config } from '../src/config'
 import { Http } from '../src/mode/cloud'
 import { HttpPure } from '../src/mode/pure'
 import { HttpSeed } from '../src/mode/seed'
