+++ packages/services-s3/src/index.ts
@@ -1,4 +1,7 @@
-// s3 injectables
+// service container modules
+export { module as s3ServicesModule } from './module'
+
+// services
 export { S3 } from './mode/cloud'
 export { S3Edge } from './mode/edge'
 export { S3Pure } from './mode/pure'
