+++ packages/services-s3/src/mode/base.ts
