+++ packages/services-s3/src/mode/edge.ts
