+++ packages/services-s3/src/mode/pure.ts
