+++ packages/testing/src/index.ts
@@ -1,1 +1,2 @@
 export * from './basic_test'
+export * from './service_test'
