+++ packages/testing/src/service_test.ts
@@ -1,15 +1,16 @@
 import { SetMode } from '@seagull/mode'
 import { Container } from 'inversify'
 
-export abstract class InjectableTest {
-  abstract injectDiModules: any[]
-  abstract inject: any[]
+export abstract class ServiceTest {
+  abstract serviceModules: any[]
+  abstract services: any[]
   injector = new Container()
 
   /** Implement your logic to run before each single test here!
    * if you override before(), you compromise internal logic of InjectableTest */
-  // tslint:disable-next-line:no-empty
-  beforeEach = () => {}
+  beforeEach() {
+    /** empty */
+  }
 
   /**
    * before every test, activate all given injectables/modules
@@ -17,10 +18,10 @@ export abstract class InjectableTest {
   before() {
     new SetMode('environment', 'pure').execute()
     this.injector = new Container()
-    for (const diMod of this.injectDiModules) {
+    for (const diMod of this.serviceModules) {
       this.injector.load(diMod)
     }
-    for (const injectable of this.inject) {
+    for (const injectable of this.services) {
       this.injector.bind(injectable).toSelf()
     }
     this.beforeEach()
@@ -29,17 +30,18 @@ export abstract class InjectableTest {
 
   /** Implement your logic to run after each single test here!
    * if you override after(), you compromise internal logic of InjectableTest */
-  // tslint:disable-next-line:no-empty
-  afterEach = () => {}
+  afterEach() {
+    /** empty */
+  }
 
   /**
    * after every test, deactivate all injectables/modules
    */
   after() {
-    for (const diMod of this.injectDiModules) {
+    for (const diMod of this.serviceModules) {
       this.injector.unload(diMod)
     }
-    for (const injectable of this.inject) {
+    for (const injectable of this.services) {
       this.injector.unbind(injectable)
     }
     this.afterEach()
